fix: move scene initialization methods to ClassesExt
